start restructuring the code to be more class-oriented and easily expandable
